Create Guess_the_number
This is a fun game and system is the opponent. System will generate the random integer number for you and gives you chance to guess it. if the digits in the number you have enter are matching and in same position as of system generated number, then its increases cows by one. if the digit in your number is in the system generated number but not in the exact position, then its increases bulls values by 1. this happens till you guess the number.